Make it optional whether a command needs to hold the lock or not

+++ apps/rush-lib/src/cli/actions/BaseRushAction.ts
@@ -12,31 +12,41 @@ import {
 import RushConfiguration from '../../data/RushConfiguration';
 import EventHooksManager from '../logic/EventHooksManager';
 
+export interface IRushCommandLineActionOptions extends ICommandLineActionOptions {
+  /** If true, no locking mechanism will be enforced when this action is run */
+  safeForSimultaneousRushProcesses?: boolean;
+}
+
 /**
  * The base Rush action that all Rush actions should extend.
  */
 export abstract class BaseRushAction extends CommandLineAction {
   private _rushConfiguration: RushConfiguration;
   private _eventHooksManager: EventHooksManager;
+  private _safeForSimultaneousRushProcesses: boolean;
 
-  constructor(options: ICommandLineActionOptions) {
+  constructor(options: IRushCommandLineActionOptions) {
     super(options);
+    this._safeForSimultaneousRushProcesses = !!options.safeForSimultaneousRushProcesses;
   }
 
   protected onExecute(): void {
     this._ensureEnvironment();
-    const lockFilePath: string = path.join(
-      this.rushConfiguration.commonTempFolder,
-      'rush.lock'
-    );
 
-    if (fsx.existsSync(lockFilePath)) {
-      fsx.removeSync(lockFilePath);
-    }
-    try {
-      fsx.openSync(lockFilePath, 'wx');
-    } catch (error) {
-      console.log(`Another rush command is already running in this repository!`);
+    if (!this._safeForSimultaneousRushProcesses) {
+      const lockFilePath: string = path.join(
+        this.rushConfiguration.commonTempFolder,
+        'rush.lock'
+      );
+      if (fsx.existsSync(lockFilePath)) {
+        fsx.removeSync(lockFilePath);
+      }
+      try {
+        fsx.openSync(lockFilePath, 'wx');
+      } catch (error) {
+        console.log(`Another rush command is already running in this repository!`);
+        process.exit(1);
+      }
     }
 
     console.log(`Starting "rush ${this.options.actionVerb}"${os.EOL}`);
